requirements.txt
pandas>=2.1.0
numpy>=1.26.0
pyarrow>=15.0.0
scikit-learn>=1.4.0
scipy>=1.11.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
plotly>=5.20.0
dash>=2.16.0
soccerdata>=1.6.0
requests>=2.31.0
httpx>=0.27.0
python-dotenv>=1.0.0
cryptography>=42.0.0
jupyter>=1.0.0
pytest>=8.0.0
